% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_translated_value}
\alias{add_translated_value}
\title{Add variables to the ABC-DS dataset located in the translated value column}
\usage{
add_translated_value(
  data,
  variables,
  translated_variables = c("examdate", "mrseqs")
)
}
\value{
A tibble or data frame
}
\description{
Add variables to the ABC-DS dataset located in the translated value column
}
\details{
Add variables to the ABC-DS dataset located in the translated value column
}
\keyword{internal}
