% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atri_pivot_wider}
\alias{atri_pivot_wider}
\title{Use a custom pivot wider that can handle an abcds_df class}
\usage{
atri_pivot_wider(data, dataset)
}
\value{
A tibble or data frame
}
\description{
Use a custom pivot wider that can handle an abcds_df class
}
\details{
Use a custom pivot wider that can handle an abcds_df class
}
\keyword{internal}
