% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{calculate_age_at_visit}
\alias{calculate_age_at_visit}
\title{Calculate Age at Visit}
\usage{
calculate_age_at_visit(data)
}
\arguments{
\item{data}{A dataset exported from the ATRI API for which to add an age_at_visit variable}
}
\value{
A data frame or \code{\link[tibble]{tibble}} with participant identifiers and
their calculated age (in years) at the time of evaluation.
}
\description{
Calculates participant age at the time of evaluation based on year of birth
and the evaluation date (\code{de_eval_date}).
}
\details{
The \code{calculate_age_at_visit()} function computes age at the time of evaluation
in years using the formula:
\deqn{age = year\_of\_evaluation - year\_of\_birth}
}
\examples{
\dontrun{
if (interactive()) {
  # Calculate age for all participants
  age_tbl <- get_demographics(de_eval_date) |> calculate_age_at_visit()

}
}

}
