% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_quarto}
\alias{format_quarto}
\title{FUNCTION_TITLE}
\usage{
format_quarto(x = NULL, type, width = 100)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
\keyword{internal}
