% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{FUNCTION_TITLE}
\usage{
get_consensus(
  ...,
  site = NULL,
  cycle = NULL,
  apply_labels = FALSE,
  controls = FALSE
)
}
\arguments{
\item{...}{One or more unquoted variable names to retrieve from the dataset.}

\item{site}{Optional; a site identifier or vector of site codes to subset data by site. Default is \code{NULL}.}

\item{cycle}{Optional; a cycle identifier or vector of cycles to subset data by cycle. Default is \code{NULL}.}

\item{apply_labels}{Logical; if \code{TRUE}, applies variable labels from the codebook to the returned data. Default is \code{FALSE}.}

\item{controls}{A boolean value that indicates whether the function should return the controls, Default is \code{FALSE}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[rlang]{defusing-advanced}}
\code{\link[tidyr]{pivot_wider}}
\code{\link[dplyr]{reexports}}
}
